[refactor] #93 - 불필요 코드 및 주석 삭제
