+++ src/main/java/com/cloudrangers/cloudpilot/CloudpilotApplication.java
@@ -1,6 +1,6 @@
 package com.cloudrangers.cloudpilot;
 
-import com.cloudrangers.cloudpilot.monitor.vcenter.config.VcenterProperties;
+import com.cloudrangers.cloudpilot.config.VcenterProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
