+++ src/main/java/com/cloudrangers/cloudpilot/config/PrometheusProperties.java
@@ -1,4 +1,4 @@
-package com.cloudrangers.cloudpilot.monitor.prometheus.config;
+package com.cloudrangers.cloudpilot.config;
 
 import lombok.Getter;
 import lombok.Setter;
