--- src/main/java/com/cloudrangers/cloudpilot/controller/AlertRuleController.java
@@ -1,18 +1,0 @@
-//package com.cloudrangers.cloudpilot.ops.api;
-//
-//@RestController
-//@RequestMapping("/ops/v1/alert-rules")
-//public class AlertRuleController {
-//
-//    @GetMapping
-//    public ApiResponse<List<AlertRuleDto>> getAll() {
-//        return ApiResponse.ok(List.of()); // 일단 빈 리스트
-//    }
-//
-//    @GetMapping("/vm/{vmId}")
-//    public ApiResponse<List<AlertRuleDto>> byVm(@PathVariable Long vmId) {
-//        return ApiResponse.ok(List.of());
-//    }
-//
-//    // create/update/delete/toggle 도 일단 "not implemented" 메시지 정도로
-//}
