+++ src/main/java/com/cloudrangers/cloudpilot/dto/HostResourceChartResponse.java
@@ -1,4 +1,4 @@
-package com.cloudrangers.cloudpilot.monitor.prometheus.dto;
+package com.cloudrangers.cloudpilot.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
