+++ src/main/java/com/cloudrangers/cloudpilot/dto/VcenterVmInfoDto.java
@@ -1,4 +1,4 @@
-package com.cloudrangers.cloudpilot.monitor.vcenter.dto;
+package com.cloudrangers.cloudpilot.dto;
 
 import lombok.Builder;
 import lombok.Value;
