+++ src/main/java/com/cloudrangers/cloudpilot/exception/ApiExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.cloudrangers.cloudpilot.common.exception;
+package com.cloudrangers.cloudpilot.exception;
 
 import com.cloudrangers.cloudpilot.common.ApiResponse;
 import org.springframework.http.ResponseEntity;
