+++ src/main/java/com/cloudrangers/cloudpilot/service/VcenterMonitorServiceImpl.java
@@ -1,8 +1,7 @@
-package com.cloudrangers.cloudpilot.monitor.vcenter.service;
+package com.cloudrangers.cloudpilot.service;
 
-import com.cloudrangers.cloudpilot.infra.vcenter.VCenterClient;
-import com.cloudrangers.cloudpilot.monitor.vcenter.dto.VcenterSummaryResponse;
-import com.cloudrangers.cloudpilot.monitor.vcenter.dto.VcenterVmInfoDto;
+import com.cloudrangers.cloudpilot.dto.VcenterSummaryResponse;
+import com.cloudrangers.cloudpilot.dto.VcenterVmInfoDto;
 import com.cloudrangers.cloudpilot.service.monitoring.PrometheusMonitoringService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
