+++ src/main/java/com/cloudrangers/cloudpilot/service/catalog/NetworkQueryService.java
@@ -1,7 +1,6 @@
 package com.cloudrangers.cloudpilot.service.catalog;
 
-import com.cloudrangers.cloudpilot.common.ApiResponse;
-import com.cloudrangers.cloudpilot.common.exception.CatalogNotFoundException;
+import com.cloudrangers.cloudpilot.exception.CatalogNotFoundException;
 import com.cloudrangers.cloudpilot.domain.catalog.NetworkSubnet;
 import com.cloudrangers.cloudpilot.dto.common.PageResponse;
 import com.cloudrangers.cloudpilot.dto.response.NetworkResponse;
