+++ src/main/java/com/cloudrangers/cloudpilot/service/monitoring/AdminOverviewServiceImpl.java
@@ -4,8 +4,8 @@ package com.cloudrangers.cloudpilot.service.monitoring;
 import com.cloudrangers.cloudpilot.dto.monitoring.AdminOverviewDto;
 import com.cloudrangers.cloudpilot.dto.monitoring.PrometheusSummaryResponse;
 import com.cloudrangers.cloudpilot.domain.vm.VmInstance;
-import com.cloudrangers.cloudpilot.monitor.prometheus.dto.VmMetricSummaryDto;
-import com.cloudrangers.cloudpilot.monitor.prometheus.service.PrometheusMetricsService;
+import com.cloudrangers.cloudpilot.dto.VmMetricSummaryDto;
+import com.cloudrangers.cloudpilot.service.PrometheusMetricsService;
 import com.cloudrangers.cloudpilot.repository.user.LoginHistoryRepository;
 import com.cloudrangers.cloudpilot.repository.vm.VmInstanceRepository;
 import lombok.RequiredArgsConstructor;
